requirements.txt
Mastodon.py == 1.8.1
opennsfw2
python-dotenv
